// fe/src/lib/api/generated/model/customerFeedbackRatingValue.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * COMPASS API
 * The version-one backend API for COMPASS.
 * OpenAPI spec version: 1.0.0
 */

export type CustomerFeedbackRatingValue = typeof CustomerFeedbackRatingValue[keyof typeof CustomerFeedbackRatingValue];


export const CustomerFeedbackRatingValue = {
  NUMBER_1: 1,
  NUMBER_2: 2,
  NUMBER_3: 3,
  NUMBER_4: 4,
  NUMBER_5: 5,
} as const;