// fe/src/lib/api/generated/model/customerFeedbackServiceValue.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * COMPASS API
 * The version-one backend API for COMPASS.
 * OpenAPI spec version: 1.0.0
 */

export type CustomerFeedbackServiceValue = typeof CustomerFeedbackServiceValue[keyof typeof CustomerFeedbackServiceValue];


export const CustomerFeedbackServiceValue = {
  COUNSELING: 'COUNSELING',
  ADMISSION: 'ADMISSION',
  TESTING: 'TESTING',
  EDUCATIONAL_INFORMATION: 'EDUCATIONAL_INFORMATION',
  REQUEST_FOR_CERTIFICATION: 'REQUEST_FOR_CERTIFICATION',
  APPLICATION_FOR_ADMISSION_TEST: 'APPLICATION_FOR_ADMISSION_TEST',
  OTHER: 'OTHER',
} as const;