// fe/src/lib/api/generated/model/customerFeedbackSubmitRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * COMPASS API
 * The version-one backend API for COMPASS.
 * OpenAPI spec version: 1.0.0
 */
import type { CustomerFeedbackAccommodatedByValue } from './customerFeedbackAccommodatedByValue';
import type { CustomerFeedbackRatingValue } from './customerFeedbackRatingValue';
import type { CustomerFeedbackServiceValue } from './customerFeedbackServiceValue';

export interface CustomerFeedbackSubmitRequest {
  accommodated_by?: CustomerFeedbackAccommodatedByValue | null;
  additional_feedback?: string;
  address?: string;
  course_year: string;
  future_service_improvement?: string;
  mobile_number?: string;
  office_cleanliness_rating: CustomerFeedbackRatingValue;
  office_environment_rating: CustomerFeedbackRatingValue;
  office_hours_rating: CustomerFeedbackRatingValue;
  office_location_rating: CustomerFeedbackRatingValue;
  /**
     * @minimum 1
     * @maximum 10000
     */
  office_visit_count: number;
  other_service?: string;
  overall_satisfaction_rating: CustomerFeedbackRatingValue;
  personnel_accommodating_rating: CustomerFeedbackRatingValue;
  personnel_appearance_rating: CustomerFeedbackRatingValue;
  personnel_availability_rating: CustomerFeedbackRatingValue;
  personnel_commitment_delivery_rating: CustomerFeedbackRatingValue;
  personnel_flexibility_rating: CustomerFeedbackRatingValue;
  personnel_information_accuracy_rating: CustomerFeedbackRatingValue;
  personnel_job_knowledge_rating: CustomerFeedbackRatingValue;
  respondent_name?: string;
  services_received: CustomerFeedbackServiceValue[];
  talked_to_guidance_counselor: boolean;
  transaction_duration: string;
}